Add keystone oidc tests
This adds tests to test getting a token (scoped and unscoped) when keystone
is configured to use oidc for authentication.  The oidc provider is keycloak.
This is based in very large part on Kristi's work in [1] and [2].

This is not quite working yet, but is close enough that we can start the
review process.  This is related to and requires changes in the dependent
keystone patch to set up the gate and keycloak server.

[1] https://github.com/knikolla/devstack-plugin-oidc
[2] https://github.com/CCI-MOC/onboarding-tools

Co-Authored-By: David Wilde <dwilde@redhat.com>
Change-Id: I1772b65f1cc3830ac293a800a79d044a6ab69d65